fix: Allow empty tables as field types
problem: annotation.rs::parse_table function crash if a field is defined
as {}.

cause: The `parse_table` was calling unwrap on an option. However if
`{}` is used as a field type, this call return None, causing a crash.

solution: Since {} is a valid (but empty) table literal, we return an
empty Vec in parse_table and generate "{}" in `Type::format_with_link()`
when we encounter an empty table.